use docker cargo config file for android builds

+++ scripts/gitlab/build-unix.sh
@@ -5,16 +5,23 @@ set -u # treat unset variables as error
 
 echo "__________Show ENVIROMENT__________"
 echo "CI_SERVER_NAME:   " $CI_SERVER_NAME
+echo "CARGO_HOME:       " $CARGO_HOME
 echo "CARGO_TARGET:     " $CARGO_TARGET
 echo "CC:               " $CC
 echo "CXX:              " $CXX
 
 echo "__________CARGO CONFIG__________"
-mkdir -p .cargo
-rm -f .cargo/config
-echo "[target.$CARGO_TARGET]" >> .cargo/config
-echo "linker= \"$CC\"" >> .cargo/config
-cat .cargo/config
+if [ "${CARGO_TARGET}" = "armv7-linux-androideabi" ]
+then
+  # use build container's cargo config
+  cat /.cargo/config
+else
+  mkdir -p .cargo
+  rm -f .cargo/config
+  echo "[target.$CARGO_TARGET]" >> .cargo/config
+  echo "linker= \"$CC\"" >> .cargo/config
+  cat .cargo/config
+fi
 
 echo "_____ Building target: "$CARGO_TARGET" _____"
 time cargo build --target $CARGO_TARGET --release --features final
